+++ app/src/main/java/thuan/todolist/feature_todo/ui/home/components/adapter/case_todo/todo_selectiontracker/ToDosKeyProvider.kt
@@ -5,11 +5,11 @@ import thuan.todolist.feature_todo.ui.home.components.adapter.case_todo.ToDoAdap
 
 class ToDosKeyProvider(private val adapter: ToDoAdapter) : ItemKeyProvider<Long>(SCOPE_CACHED) {
     override fun getKey(position: Int): Long {
-        return adapter.dataList[position].id.toLong()
+        return adapter.dataList[position].id
     }
 
     override fun getPosition(key: Long): Int {
         // indexOfFirst returns the first index of the element for which the given predicate is true.
-        return adapter.dataList.indexOfFirst { it.id.toLong() == key }
+        return adapter.dataList.indexOfFirst { it.id == key }
     }
 }
